chore: Track the internal error to see how many users are affected [EXT-263]
## Description

There is a [bug in Chromium](https://issues.chromium.org/issues/367440970) that causes an `Internal error` when we try to enable a ruleset.

We want to know how many users are affected by this error as well the ruleset ID that is affected to check if there are hints into how to solve the error. 

AdBlock will send an event to its servers with the error message which will contain the ruleset ID, as part of [this MR in the WebExt SDK](https://gitlab.com/eyeo/adblockplus/abc/webext-ad-filtering-solution/-/merge_requests/1142)

## Motivation and Context

- We got reports of ads being visible
- AAPVs are dropping
- Premium campaigns are not triggered

## Checklist:

- [x] My changes didn't require a documentation update
- [x] My changes didn't require new tests
- [x] All new and existing tests passed


## Depends on:
- [x] https://gitlab.com/eyeo/adblockplus/abc/webext-ad-filtering-solution/-/merge_requests/1142

Closes EXT-263

See eyeo/extensions/extensions!250