fix: do not wait for fetch to return [EXT-587]
## Description
The changes in this MR remove the "wait" for the telemetry fetch to be complete. This change will remove the latency in waiting for the fetch request to complete, reducing the chance that the service worker will be shut down before calculating the next ping time. 

Note that no data is returned to the extension from the telemetry server.


## Motivation and Context
After the migration to MV3, there has been a significant drop in telemetry data from our MV3 users. We’ve made a few changes to the Telemetry servers and to the AdBlock telemetry code to determine if there’s an issue with the code.  After reviewing this error data and getting feedback from the Chrome developers, the next step is this change to the AdBlock telemetry code.


## Checklist:

<!-- Go over all the following points, keep those that apply and mark them with an `x` whenever the task is completed -->

- [X] I have updated the release notes
- [X] My changes didn't require a documentation update
- [X] My changes didn't require new tests
- [X] All new and existing tests passed

See eyeo/extensions/extensions!263