feat: Adblock Add state to Options Page, Part I [EXT-420]
## Description

This MR adds the conditional state check for checkboxes on the options page and reorganizes the components from !209 to better separate data for a specific instance of the list and a general options list. 

## Motivation and Context

Due to planning changes, we will be adding the first new options not to the general tab, but to the filters tab. This MR collects the work done for the general tab to this point and abstracts it so that the list may be used on the filters tab. 

## Checklist:
- [x] My changes didn't require a documentation update
- [x] My changes didn't require new tests
- [ ] All new and existing tests passed

See eyeo/extensions/extensions!274