build: Quick review of the repo import script to bring it up to date [noissue]
## Description

- Primarily doc fixes. Updates comments to bring them up to date with the current reality.
- Fixes one bug introduced when we removed creating gitlab child pipelines from this script.

## Motivation and Context

This script may potentially be used to import the webext SDK into the monorepo in the coming months. While reading over it in preparation for a handover session, I noticed that the comments were out of date. Some referred to functionality that had been later removed from the script.

## Checklist:

- [x] I have updated the release notes
- [x] My changes require a documentation update and I have updated the documentation accordingly
- [x] My changes didn't require new tests

See eyeo/extensions/extensions!278