feat: Add an test split experiment to AdBlock [EXT-612]
## Description

This MR creates a test split experiment, which logs different messages to AdBlock's telemetry based on which variant a client is assigned to.

We should be able too see the spread in the data following the experiment config from https://easylist-downloads.adblockplus.org/ab-testing/experiments.json.

## Motivation and Context

This is to verify that split experiments create the splits that we expect in the real world.

## Checklist:

<!-- Go over all the following points, keep those that apply and mark them with an `x` whenever the task is completed -->

- [x] I have updated the release notes
- [x] My changes didn't require a documentation update
- [x] My changes didn't require new tests
- [x] All new and existing tests passed

See eyeo/extensions/extensions!282