test: Use wdio polyfill on all ABP tests [Refs EXT-608]
<!-- Provide a general summary of your changes in the Title above -->
<!-- If this is a draft: add "Draft:" to the Title. -->

## Description

Uses wdio polyfill on all ABP tests:

- A new `host/adblockplus/test/end-to-end/tests/index.mjs` file defines the non-upgrade test suites.
- Adds / updates polyfill functions whenever needed
- Updates tests code to make it work for selenium
- Removes the SUITE env variable and `--suite` npm parameter
- Several test cases get skipped due to time development constraints. They will be fixed and unskipped on an upcoming MR.

To be done on a separate MR:

- add new `runner.uninstallCustom.mjs`. In this MR the "uninstall custom" test won't run.
- remove unused WDIO packages

This MR is still a draft because some test cases are failing on the CI. Please feel free to make changes in this MR in order to fix them.

## Motivation and Context

The reason for the issue is unclear, since the nightly run passed  and no new commits have landed to the main branch.

In any case, switching the non-upgrade ABP tests to run selenium using the polyfill fixes the issue.

## Checklist:

<!-- Go over all the following points, keep those that apply and mark them with an `x` whenever the task is completed -->

- [X] My changes didn't require a documentation update
- [X] My changes didn't require new tests
- [ ] All new and existing tests passed

See eyeo/extensions/extensions!277