docs: Document our current Opera store notes to reviewers for AdBlock [noissue]
## Description

This updates our notes for reviewers for AdBlock to have a section for the Opera store. This is a copy of what is already on the Opera store, which in turn appears to be based on Adblock Plus's simpler notes for reviewers.

## Motivation and Context

The Opera store listing for AdBlock is quite new, and it seems that this documentation was missed.

## Checklist:

- [x] I have updated the release notes
- [x] My changes require a documentation update and I have updated the documentation accordingly
- [x] My changes didn't require new tests
- [x] All new and existing tests passed

See eyeo/extensions/extensions!287