test: Move blocking-hiding-testpage to localhost, unskip Public API tests on ABP [Refs EXT-608, Refs EXT-603]
<!-- Provide a general summary of your changes in the Title above -->
<!-- If this is a draft: add "Draft:" to the Title. -->

## Description

- Copies `blocking-hiding-testpage.html` to the local test server, adapting test cases in both hosts accordingly
- Unskips `Public API` tests on ABP on Chromium and Firefox
- Adds minor improvements to the E2E test code

## Motivation and Context

This is a follow up on !277, which had to skip several test cases. The unskipped tests in this MR use the `blocking-hiding-testpage.html` page, so this MR takes the chance to copy that page to the local test server. Optimizing that page is out of scope, to be done by https://eyeo.atlassian.net/browse/EXT-603

## Checklist:

- [X] My changes didn't require a documentation update
- [X] My changes didn't require new tests
- [X] All new and existing tests passed

See eyeo/extensions/extensions!289