feat: Add inline toggle variant [noissue]
## Description

| in storybook | in the app | current |
|--------------|------------|---------|
| ![Screenshot_2024-11-26_at_13.03.23](/uploads/1baa0d0d65a46ec74f2a43a208b4cbb6/Screenshot_2024-11-26_at_13.03.23.png)             | ![Screenshot_2024-11-26_at_13.10.51](/uploads/59bc9a365cb0ec7257647e1fad8550b0/Screenshot_2024-11-26_at_13.10.51.png)            |   ![Screenshot_2024-11-26_at_13.10.40](/uploads/ea2ff18f2c839aa41bd99ead3cc91ba7/Screenshot_2024-11-26_at_13.10.40.png)      |

## Motivation and Context

As mentioned in !209, I have concerns about React changes being appropriate for Adblock without the inline-styled toggle. This adds it, with the same behavior as currently exists. Again, I anticipate that when we handle redesigns in the new year, this will change, so I have tried to match behavior without spending too long perfecting it. :grin: 

## Checklist:
- [x] My changes didn't require a documentation update
- [x] I have added tests to cover my changes
- [x] All new and existing tests passed

See eyeo/extensions/extensions!290