test: Fix has premium features timeout issues (AB) [noissue]
## Description

Fixes timeout issues on the `has premium features` test (AB) by polling the button click to handle a sliding animation.

## Motivation and Context

This is a follow up on !305, which had a side effect of creating flakiness to the `has premium features` test.

I was able to reproduce the issue locally on a regular E2E `npm run`, and verified the fix running it 10 times in a row with:

```
for i in {1..10}; do BROWSER=chromium MANIFEST_VERSION=3 npm run --workspace host/adblock test:end-to-end-local -- -g "has premium features"; done
```

## Checklist:

- [X] My changes didn't require a documentation update
- [X] My changes didn't require new tests
- [x] All new and existing tests passed

See eyeo/extensions/extensions!310