test: Add eyeometry test [EXT-507]
We need to make sure telemetry has sent the data and saved the last ping date.

## Description

In this MR we're waiting for 10 seconds and check the storage to make sure the telemetry saved the last ping date.
Both Adblock and ABP are covered.

Eyeometry tests depend on env vars like `EYEOMETRY_BEARER` to be set at extension build time. A `before` hook handles those checks, which are strict on the CI and relaxed on the local run.

## Checklist:

- [x] My changes didn't require a documentation update
- [x] New tests have been added
- [x] All new and existing tests passed

See eyeo/extensions/extensions!215