test: Increase the timeout for clickAndCloseNewTab elements [noissue]
## Description

Increases the timeout for `clickAndCloseNewTab` elements.

## Motivation and Context

Flaky job example: https://gitlab.com/eyeo/extensions/extensions/-/jobs/8540414235

![shows_image_swap_off](/uploads/0ca8c5521a576081f195dfb63e68f1ad/shows_image_swap_off.png)

I wasn't able to reproduce the issue locally, but as seen in the screenshot, the `Get it now` button didn't finish the animation to show up in the header, that's why the click failed. Increasing the timeout on the header elements should fix that flakiness.

## Checklist:

- [X] My changes didn't require a documentation update
- [X] My changes didn't require new tests
- [X] All new and existing tests passed

See eyeo/extensions/extensions!316