build: Automate the release step of merging the release branch back into main [EXT-549]
## Description

This adds a new "create_release_mr_back_to_main" job to the pipelines for both extensions. This job is run in the same situation as automatically creating a gitlab release: on the pipelines for a release tag.

This job creates a merge request in the extensions repo to bring the changes in the release branch (version bump and updating the release notes) back into the main branch.

After this has been merged, the [Update the main branch](https://gitlab.com/r.deysel/extensions/-/merge_requests/10) section of the release process will be changed to only be approving and merging this automatically created merge request.

## Motivation and Context

Creating this MR is currently a manual step in our release process.

## Checklist:

- [ ] My changes require a documentation update and I have updated the documentation accordingly
  - The release process will be updated accordingly after this MR is merged.
- [x] My changes didn't require new tests
  - Changes to the pipelines themselves are difficult to test, but I did test this manually using a fork of the extensions repo. See https://gitlab.com/r.deysel/extensions/-/merge_requests/10 as an example MR created by this new job.
- [x] All new and existing tests passed

See eyeo/extensions/extensions!314