test: Remove 1 minute waits from tests [EXT-181]
## Description

We had 2 test in ABP that stopped test execution for 1 minute in order to ensure certain text will be displayed before test is run. 

The first of those tests, `should update a single filter list`, was removed, because it was agreed it is not a crucial test and we believe the majority of users will just update all filter lists, instead of a single one.

The second test, `should update all filter lists` was changed so that instead of pausing the test execution for 1 minute, we now periodically check for certain text to be displayed. 

The order of `Advanced Tab - Filter Lists` tests and Ad Filtering tests was rearranged to maximize the probability that more than 1 minute will pass before `should update all filter lists` test is run, and to limit any other filterlist changes done before that (like removing and/or disabling them)

## Checklist:

- [x] My changes didn't require a documentation update
- [x] My changes didn't require new tests
- [x] All new and existing tests passed

See eyeo/extensions/extensions!325