test: moves re-adding AA filterlist to after sequence [noissue]
## Description

This MR reverts the changes that were introduced to the `ad-filtering.js` test suite in: https://gitlab.com/eyeo/extensions/extensions/-/merge_requests/325/diffs#21a750ba1d0c928081898f8c88bde9f6637c2e79

## Motivation and Context

Part of the Acceptable Ads E2E smoke test is that the AA filterlist is removed. Because subsequent tests rely on the AA filterlist to be present, it was added back in the after sequence. 

In https://gitlab.com/eyeo/extensions/extensions/-/merge_requests/325 we moved the re-adding of the AA filterlist from the after sequence to the AA test itself, in order to speed up test execution (see this comment for more context: https://gitlab.com/eyeo/extensions/extensions/-/merge_requests/325#note_2248564940).

This turned out to be a bad idea, because if the Acceptable Ads test fails, the filterlist is not added back and any subsequent tests that rely on that filter being there, will also fail.

## Checklist:

- [x] My changes didn't require a documentation update
- [x] My changes didn't require new tests
- [ ] All new and existing tests passed

See eyeo/extensions/extensions!329